Fixed the typo. Improved the warning messages.
